fix: account for parent classes as well
If a Schema Interface is composed of several other parent classes,
their fields were not getting a default value. i.e.

class MyOtherInterface(Schema):
    location = schema.TextLine()

class PrimaryInterface(Schema, MyOtherInterface):
    zip_code = schema.TextLine()

If PrimaryInterface is the main schema of a Content Type or a behavior,
the `initialize_missing_attributes` function would not find out that
there is a `location` field on that interface.